requirements.txt
asgiref==3.9.1
Django==5.2.4
django-filter==25.1
django-jazzmin==3.0.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
